requirements.txt
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
gunicorn==20.1.0
psycopg2-binary
whitenoise==6.2.0
dj-database-url